moved comment to separate file and fixed typo in duns_verification
